Refactor RADAR export
